tests-hil: split the scenario drivers into a shared module
The Net fixture wrapper and the join_ok/join_never/join_dlsettings
drivers move to tests/common so a second DUT's suite can register the
same scenarios against its own firmware crate. No behavior change.
